requirements.txt
Flask==3.0.0
flask-cors==4.0.0
openai-whisper==20231117
# 降低numpy版本以确保与librosa等库兼容
numpy==1.23.5
soundfile==0.13.0
pydub==0.25.1
python-dotenv==1.0.0
requests==2.31.0
setuptools==68.2.2

# 说话人识别相关
pyannote.audio==3.1.1
speechbrain==0.5.16
resemblyzer==0.1.3
scikit-learn==1.3.0
scipy==1.11.2
huggingface_hub==0.20.3

# 音频处理
librosa==0.10.1
webrtcvad==2.0.10

torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2